Fixing stuff, added inital team management view, changed grid into flex containers, dividing code in separate files and folders
