+++ frontend_ta/src/components/DisplayTournament.js
@@ -1,0 +1,9 @@
+import '../styles/App.css';
+
+export function DisplayTournament() {
+    return (
+      <div class="main-panel">
+        <h2>List of tournaments</h2>
+      </div>
+    );
+  }
