+++ frontend_ta/src/components/signUp.js
@@ -1,8 +1,7 @@
 import React from 'react';
-import postData from './utils';
-import SignUp from './signUp';
-import "./SignInUp.css";
-class SignIn extends React.Component {
+import SignIn from './SignIn';
+import postData from '../utils';
+class SignUp extends React.Component {
     constructor(props) {
         super(props);
         this.state = {
@@ -10,7 +9,9 @@ class SignIn extends React.Component {
           isSubmitted:false,
           username :'',
           password :'',
-          view :"SignIn"
+          confirmpassword :'',
+          email:'',
+          view :"SignUp"
         };
         this.handleChange = this.handleChange.bind(this);
         this.handleSubmit = this.handleSubmit.bind(this);
@@ -26,11 +27,14 @@ class SignIn extends React.Component {
  handleChange(event) {
   if(event.target.name=="username")this.setState({username: event.target.value});
   else if (event.target.name=="password") this.setState({password: event.target.value});
+  else if (event.target.name=="email") this.setState({email: event.target.value});
+  else if (event.target.name=="confirmpassword") this.setState({confirmpassword: event.target.value});
+
  }
 
   handleSubmit(event) {
     event.preventDefault();
-    let data={username: this.state.username,password:this.state.password};
+    let data={username: this.state.username,password:this.state.password,email:this.state.email,confirmpassword:this.state.confirmpassword};
     console.log(data);
     let url=""
     /*
@@ -43,13 +47,11 @@ class SignIn extends React.Component {
 
     
    render() {
-    let nextcomponent;
-    console.log( "la view è"  + this.state.view);
-    if (this.state.view=="SignUp") nextcomponent= (<SignUp/>);
-    if(this.state.view=="SignIn") nextcomponent= (
-      <div className="app">
-        <div className="login-form">
-          <div className="title">Sign In</div>
+    if (this.state.view=="SignIn") return (<SignIn/>);
+    else if(this.state.view=="SignUp") return (
+<div className="app">
+    <div className="login-form">
+          <div className="title">Sign Up</div>
     <div className="form">
       <form onSubmit={this.handleSubmit}>
         <div className="input-container">
@@ -57,25 +59,27 @@ class SignIn extends React.Component {
           <input type="text" name="username" value={this.state.username} onChange={this.handleChange} required />
         </div>
         <div className="input-container">
+          <label>Email </label>
+          <input type="text" name="email" value={this.state.email} onChange={this.handleChange} required />
+        </div>
+        <div className="input-container">
           <label>Password </label>
           <input type="password" name="password" value={this.state.password} onChange={this.handleChange} required />
         </div>
+         <div className="input-container">
+          <label>Confirm Password </label>
+          <input type="password" name="confirmpassword" value={this.state.confirmpassword} onChange={this.handleChange} required />
+         </div>
         <div className="button-container">
-          <input type="submit" value="Sign in"/>
+          <input type="submit" value="Sign up"/>
         </div>
       </form>
       {this.renderErrorMessage()}
-      <button className="formbutton" onClick={()=>this.setState({view:"SignUp"})}> Create Account </button>
-      
+      <button className="formbutton" onClick={()=>this.setState({view:"SignIn"})}> Sign in if you already have an account </button>
     </div>
   </div>
 </div>
-
-    
  );
-     return nextcomponent;
     }
-    
   }
-  
-export default SignIn;
+export default SignUp;
