+++ frontend_ta/src/index.js
@@ -1,14 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './styles/index.css';
 import reportWebVitals from './reportWebVitals';
-import Login from './signIn';
+import App from './App' 
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
-  <Login />
-  
+    <App />
 );
 
 // If you want to start measuring performance in your app, pass a function
