+++ frontend_ta/src/styles/SignInUp.css
@@ -10,7 +10,6 @@
     justify-content: left;
   }
   
-  
   input[type="text"],
   input[type="password"] {
     height: 25px;
@@ -18,19 +17,7 @@
     width: 250px;
   }
   
-  input[type="submit"] {
-    margin-top: 10px;
-    cursor: pointer;
-    font-size: 15px;
-    background: #01d28e;
-    border: 1px solid #01d28e;
-    color: #fff;
-    padding: 10px 20px;
-  }
-  
-  input[type="submit"]:hover {
-    background: #6cf0c2;
-  }
+
   
   .button-container {
     display: flex;
@@ -46,8 +33,6 @@
     flex-direction: column;
   }
   
-  
-  
   .error {
     color: red;
     font-size: 12px;
@@ -58,10 +43,5 @@
     margin-bottom: 20px;
   }
   
-  .input-container {
-    display: flex;
-    flex-direction: column;
-    gap: 8px;
-    margin: 10px;
-  }
+
