+++ frontend_ta/src/styles/forms.css
@@ -1,0 +1,28 @@
+
+input[type="submit"] {
+    margin-top: 10px;
+    cursor: pointer;
+    font-size: 15px;
+    background: #01d28e;
+    border: 1px solid #01d28e;
+    color: #fff;
+    padding: 10px 20px;
+}
+
+input[type="submit"]:hover {
+    background: #6cf0c2;
+}
+
+.input-container {
+    display: flex;
+    flex-direction: column;
+    gap: 8px;
+    margin: 10px;
+}
+
+.selector{
+    margin-left: 40%;
+    margin-right: 40%;
+    padding: 1rem;
+    margin-bottom: 1rem;
+}
